% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_set_user_settings}
\alias{cognitoidentityprovider_admin_set_user_settings}
\title{This action is no longer supported}
\usage{
cognitoidentityprovider_admin_set_user_settings(UserPoolId, Username,
  MFAOptions)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool that contains the user that you are setting
options for.}

\item{Username}{[required] The user name of the user that you are setting options for.}

\item{MFAOptions}{[required] You can use this parameter only to set an SMS configuration that uses
SMS for delivery.}
}
\description{
\emph{This action is no longer supported.} You can use it to configure only
SMS MFA. You can't use it to configure TOTP software token MFA. To
configure either type of MFA, use
\href{https://docs.aws.amazon.com/cognito-user-identity-pools/latest/APIReference/API_AdminSetUserMFAPreference.html}{AdminSetUserMFAPreference}
instead.
}
\section{Request syntax}{
\preformatted{svc$admin_set_user_settings(
  UserPoolId = "string",
  Username = "string",
  MFAOptions = list(
    list(
      DeliveryMedium = "SMS"|"EMAIL",
      AttributeName = "string"
    )
  )
)
}
}

\keyword{internal}
