% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_global_sign_out}
\alias{cognitoidentityprovider_global_sign_out}
\title{Signs out users from all devices}
\usage{
cognitoidentityprovider_global_sign_out(AccessToken)
}
\arguments{
\item{AccessToken}{[required] The access token.}
}
\description{
Signs out users from all devices. It also invalidates all refresh tokens
issued to a user. The user's current access and Id tokens remain valid
until their expiry. Access and Id tokens expire one hour after they are
issued.
}
\section{Request syntax}{
\preformatted{svc$global_sign_out(
  AccessToken = "string"
)
}
}

\keyword{internal}
