% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_users_in_group}
\alias{cognitoidentityprovider_list_users_in_group}
\title{Lists the users in the specified group}
\usage{
cognitoidentityprovider_list_users_in_group(UserPoolId, GroupName,
  Limit, NextToken)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool.}

\item{GroupName}{[required] The name of the group.}

\item{Limit}{The limit of the request to list users.}

\item{NextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Lists the users in the specified group.
}
\details{
Calling this action requires developer credentials.
}
\section{Request syntax}{
\preformatted{svc$list_users_in_group(
  UserPoolId = "string",
  GroupName = "string",
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
