% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_service_linked_role_deletion_status}
\alias{iam_get_service_linked_role_deletion_status}
\title{Retrieves the status of your service-linked role deletion}
\usage{
iam_get_service_linked_role_deletion_status(DeletionTaskId)
}
\arguments{
\item{DeletionTaskId}{[required] The deletion task identifier. This identifier is returned by the
DeleteServiceLinkedRole operation in the format
\verb{task/aws-service-role/&lt;service-principal-name&gt;/&lt;role-name&gt;/&lt;task-uuid&gt;}.}
}
\description{
Retrieves the status of your service-linked role deletion. After you use
the DeleteServiceLinkedRole API operation to submit a service-linked
role for deletion, you can use the \code{DeletionTaskId} parameter in
\code{GetServiceLinkedRoleDeletionStatus} to check the status of the
deletion. If the deletion fails, this operation returns the reason that
it failed, if that information is returned by the service.
}
\section{Request syntax}{
\preformatted{svc$get_service_linked_role_deletion_status(
  DeletionTaskId = "string"
)
}
}

\keyword{internal}
