% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_list_saml_providers}
\alias{iam_list_saml_providers}
\title{Lists the SAML provider resource objects defined in IAM in the account}
\usage{
iam_list_saml_providers()
}
\description{
Lists the SAML provider resource objects defined in IAM in the account.
}
\details{
This operation requires \href{https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html}{Signature Version 4}.
}
\section{Request syntax}{
\preformatted{svc$list_saml_providers()
}
}

\keyword{internal}
