% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_account_assignments}
\alias{ssoadmin_list_account_assignments}
\title{Lists the assignee of the specified AWS account with the specified
permission set}
\usage{
ssoadmin_list_account_assignments(InstanceArn, AccountId,
  PermissionSetArn, MaxResults, NextToken)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the SSO instance under which the operation will be executed.
For more information about ARNs, see Amazon Resource Names (ARNs) and
AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{AccountId}{[required] The identifier of the AWS account from which to list the assignments.}

\item{PermissionSetArn}{[required] The ARN of the permission set from which to list assignments.}

\item{MaxResults}{The maximum number of results to display for the assignment.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}
}
\description{
Lists the assignee of the specified AWS account with the specified
permission set.
}
\section{Request syntax}{
\preformatted{svc$list_account_assignments(
  InstanceArn = "string",
  AccountId = "string",
  PermissionSetArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
