% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_instances}
\alias{ssoadmin_list_instances}
\title{Lists the SSO instances that the caller has access to}
\usage{
ssoadmin_list_instances(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of results to display for the instance.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}
}
\description{
Lists the SSO instances that the caller has access to.
}
\section{Request syntax}{
\preformatted{svc$list_instances(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
