% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_update_instance_access_control_attribute_configuration}
\alias{ssoadmin_update_instance_access_control_attribute_configuration}
\title{Updates the AWS SSO identity store attributes to use with the AWS SSO
instance for attributes-based access control (ABAC)}
\usage{
ssoadmin_update_instance_access_control_attribute_configuration(
  InstanceArn, InstanceAccessControlAttributeConfiguration)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the SSO instance under which the operation will be executed.}

\item{InstanceAccessControlAttributeConfiguration}{[required] Updates the attributes for your ABAC configuration.}
}
\description{
Updates the AWS SSO identity store attributes to use with the AWS SSO
instance for attributes-based access control (ABAC). When using an
external identity provider as an identity source, you can pass
attributes through the SAML assertion as an alternative to configuring
attributes from the AWS SSO identity store. If a SAML assertion passes
any of these attributes, AWS SSO will replace the attribute value with
the value from the AWS SSO identity store. For more information about
ABAC, see Attribute-Based Access Control in the \emph{AWS SSO User Guide}.
}
\section{Request syntax}{
\preformatted{svc$update_instance_access_control_attribute_configuration(
  InstanceArn = "string",
  InstanceAccessControlAttributeConfiguration = list(
    AccessControlAttributes = list(
      list(
        Key = "string",
        Value = list(
          Source = list(
            "string"
          )
        )
      )
    )
  )
)
}
}

\keyword{internal}
